Require list indentation to use spaces
This fixes a stack overflow when a list item was indented with mixed
whitespace (ex. tabs and spaces). The parser initially recognized the
item as the start of a list, but then the list parser rejected it for
inconsistent indentation. It's unclear how a tab should be interpreted
in indentation anyway (how many spaces is it equivalent to?), so KWML
will instead require lists to be indented with spaces only.